Move GetDisplayVersion to shared AssemblyVersionHelper
Consolidate duplicate GetDisplayVersion implementations from
Dashboard's AssemblyExtensions and AppHost's AuxiliaryBackchannelRpcTarget
into the shared AssemblyVersionHelper. Dashboard extension method now
delegates to the shared helper.
